+++ modules/server-node/src/index.ts
@@ -16,6 +16,7 @@ import {
   jsonifyError,
   GetTransfersFilterOpts,
   GetTransfersFilterOptsSchema,
+  VectorErrorJson,
 } from "@connext/vector-types";
 import { constructRpcRequest, getPublicIdentifierFromPublicKey, hydrateProviders } from "@connext/vector-utils";
 import { WithdrawCommitment } from "@connext/vector-contracts";
@@ -26,6 +27,11 @@ import { PrismaStore } from "./services/store";
 import { config } from "./config";
 import { createNode, deleteNodes, getChainService, getNode, getNodes } from "./helpers/nodes";
 import { ServerNodeError } from "./helpers/errors";
+import {
+  ResubmitWithdrawalResult,
+  startWithdrawalSubmissionTask,
+  submitUnsubmittedWithdrawals,
+} from "./services/withdrawal";
 
 const configuredIdentifier = getPublicIdentifierFromPublicKey(Wallet.fromMnemonic(config.mnemonic).publicKey);
 export const logger = pino({ name: configuredIdentifier, level: config.logLevel ?? "info" });
@@ -55,6 +61,10 @@ server.addHook("onReady", async () => {
     logger.info({ node: nodeIndex }, "Rehydrating persisted node");
     await createNode(nodeIndex.index, store, storedMnemonic, config.skipCheckIn ?? false);
   }
+
+  // submit all withdrawals older than a week or any mainnet withdrawals when
+  // gas price is low
+  startWithdrawalSubmissionTask(store);
 });
 
 server.get("/ping", async () => {
@@ -969,6 +979,51 @@ server.post<{ Body: NodeParams.Admin }>(
           storeError: e.message,
         }).toJson(),
       );
+    }
+  },
+);
+
+server.post<{ Body: NodeParams.SubmitWithdrawals }>(
+  "/withdraw/submit",
+  {
+    schema: {
+      body: NodeParams.SubmitWithdrawalsSchema,
+      response: NodeResponses.SubmitWithdrawalsSchema,
+    },
+  },
+  async (request, reply) => {
+    if (request.body.adminToken !== config.adminToken) {
+      return reply
+        .status(401)
+        .send(new ServerNodeError(ServerNodeError.reasons.Unauthorized, "", request.body).toJson());
+    }
+    try {
+      const nodes = getNodes();
+      const channels = await store.getChannelStates();
+      const results: { [identifer: string]: ResubmitWithdrawalResult[] | VectorErrorJson } = {};
+      for (const node of nodes) {
+        // gather all unsubmitted withdrawal commitments for all channels
+        const nodeChannels = channels.filter((chan) => chan.aliceIdentifier === node.node.publicIdentifier);
+        const nodeResults = await submitUnsubmittedWithdrawals(nodeChannels, store);
+        if (nodeResults.isError) {
+          logger.error(
+            { error: jsonifyError(nodeResults.getError()!), publicIdentifier: node.node.publicIdentifier },
+            "Failed to submit withdrawals",
+          );
+        } else {
+          logger.info(
+            { results: nodeResults.getValue(), publicIdentifier: node.node.publicIdentifier },
+            "Submitted withdrawals for node",
+          );
+        }
+        results[node.node.publicIdentifier] = nodeResults.isError
+          ? jsonifyError(nodeResults.getError()!)
+          : nodeResults.getValue();
+      }
+      logger.info({ nodes: nodes.map((n) => n.node.publicIdentifier), results }, "Completed withdrawal submission");
+      return reply.status(200).send(results);
+    } catch (e) {
+      return reply.status(500).send(jsonifyError(e));
     }
   },
 );